+++ apps/phoenix_duskmoon/lib/phoenix_duskmoon.ex
@@ -2,7 +2,7 @@ defmodule PhoenixDuskmoon do
   @moduledoc """
   Provides Duskmoon UI for Phoenix project.
 
-  Require `tailwindcss >= 4.0` and `daisyui >= 5.0`
+  Require `tailwindcss >= 4.0` and `duskmoonui >= 5.0`
 
   ## Install in deps
 
@@ -36,7 +36,7 @@ defmodule PhoenixDuskmoon do
 
       @import "tailwindcss";
       @plugin "@tailwindcss/typography";
-      @plugin "daisyui";
+      @plugin "duskmoonui";
       @import "phoenix_duskmoon/theme";
       @import "phoenix_duskmoon/components";
   ```